pyproject.toml
[project]
name = "sit-wechat-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "pydantic>=2.10.6",
    "python-json-logger>=3.3.0",
    "pyyaml>=6.0.2",
    "wcferry>=39.4.2.2",
]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
package = true

[project.scripts]
wcferry-agent = "wcferry_agent:main"
wcferry-mcp-server = "wcferry_mcp_server:main"
wcferry-message-handler = "wcferry_message_handler:main"
